sendmail.in: Handle mmda exit codes, send mail in case of errors.
sendmail now handles the various non-zero mmda exit codes according
to the error type.

If a recipient has errors in .forward processing,
a mail is sent to the recipient notifying about the error. This mail
is sent with forced mbox delivery.

If some local or external addresses could not be delivered into, a mail
is sent to the sender about it.